chore: proper pagination and question
