+++ package.json
@@ -43,6 +43,7 @@
     "@aws-sdk/s3-request-presigner": "3.485.0",
     "@chakra-ui/react": "^2.8.2",
     "@hookform/resolvers": "^3.3.4",
+    "@nikolovlazar/chakra-ui-prose": "^1.2.1",
     "@opengovsg/design-system-react": "^1.15.0",
     "@opengovsg/sgid-client": "^2.2.0",
     "@paralleldrive/cuid2": "^2.2.2",
