+++ src/server/modules/watson/watson.router.ts
@@ -66,6 +66,7 @@ export const watsonRouter = router({
             sqlQuery: true,
             badResponseReason: true,
             isGoodResponse: true,
+            question: true,
             id: true,
             createdAt: true,
           },
@@ -140,6 +141,8 @@ export const watsonRouter = router({
       }),
     )
     .query(async ({ ctx: { user }, input: { messageId, limit, offset } }) => {
+      await new Promise((res) => setTimeout(res, 2000))
+
       const { sqlQuery } = await prisma.chatMessage.findFirstOrThrow({
         where: { id: messageId, conversation: { userId: user.id } },
         select: {
