+++ src/theme/index.ts
@@ -4,13 +4,18 @@ import { shadows } from './foundations/shadows'
 import { layerStyles } from './layerStyles'
 import { components } from './components'
 import { textStyles } from './foundations/textStyles'
+import { withProse } from '@nikolovlazar/chakra-ui-prose'
 
-export const theme = extendTheme(ogpDsTheme, {
-  shadows,
-  components: {
-    ...ogpDsTheme.components,
-    ...components,
+export const theme = extendTheme(
+  ogpDsTheme,
+  {
+    shadows,
+    components: {
+      ...ogpDsTheme.components,
+      ...components,
+    },
+    textStyles,
+    layerStyles,
   },
-  textStyles,
-  layerStyles,
-})
+  withProse(),
+)
